+++ src/main/java/nextstep/subway/applicaion/dto/LineDto.java
@@ -5,7 +5,6 @@ import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import nextstep.subway.domain.entity.Line;
 
-import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -31,7 +30,6 @@ public class LineDto {
                 .stations(
                         line.getStations().stream()
                                 .map(StationDto::of)
-                                .sorted(Comparator.comparing(StationDto::getId))
                                 .collect(Collectors.toUnmodifiableList()))
                 .build();
     }
